# openapi.yaml
openapi: 3.0.3
info:
  title: QAIra API
  version: 1.0.0
  description: |
    OpenAPI specification for the current QAIra Fastify + PostgreSQL backend.
    This spec documents the implemented CRUD resources, request payloads,
    query filters, and the execution lifecycle actions.
servers:
  - url: http://localhost:3000
    description: Local development server
...
paths:
...
  /test-suites:
    get:
      tags: [Test Suites]
      summary: List test suites
      parameters:
        - $ref: "#/components/parameters/AppTypeIdQuery"
        - $ref: "#/components/parameters/ParentIdQuery"
      responses:
        "200":
          description: List of test suites
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TestSuite"
    post:
      tags: [Test Suites]
      summary: Create test suite
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTestSuiteRequest"
      responses:
        "200":
          description: Test suite created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdResponse"
  /test-suites/{id}:
    parameters:
      - $ref: "#/components/parameters/IdParam"
    get:
      tags: [Test Suites]
      summary: Get test suite by ID
      responses:
        "200":
          description: Test suite record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestSuite"
    put:
      tags: [Test Suites]
      summary: Update test suite
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTestSuiteRequest"
      responses:
        "200":
          description: Update result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdatedResponse"
    delete:
      tags: [Test Suites]
      summary: Delete test suite
      responses:
        "200":
          description: Delete result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeletedResponse"
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
...
    AppTypeIdQuery:
      name: app_type_id
      in: query
      schema:
        type: string
    ParentIdQuery:
      name: parent_id
      in: query
      schema:
        type: string
        description: Use the literal string `null` to filter root suites.
...
  schemas:
    IdResponse:
      type: object
      properties:
        id:
          type: string
      required: [id]
    UpdatedResponse:
      type: object
      properties:
        updated:
          type: boolean
          example: true
      required: [updated]
    DeletedResponse:
      type: object
      properties:
        deleted:
          type: boolean
          example: true
      required: [deleted]
...
    TestSuite:
      type: object
      properties:
        id: { type: string }
        app_type_id: { type: string }
        name: { type: string }
        parent_id: { type: string, nullable: true }
        created_at: { type: string }
    CreateTestSuiteRequest:
      type: object
      required: [app_type_id, name]
      properties:
        app_type_id: { type: string }
        name: { type: string }
        parent_id: { type: string }
    UpdateTestSuiteRequest:
      type: object
      properties:
        name: { type: string }
        parent_id: { type: string }
...